Move interfaces to struct files
