+++ src/data/game_table.go
@@ -2,18 +2,30 @@ package data
 
 import (
 	"math/rand"
-
-	i "github.com/holy-tech/discord-roulette/src/interfaces"
 )
 
+//go:generate mockgen --destination=./../../mocks/table.go github.com/holy-tech/discord-roulette/src/data Table
+type Table interface {
+	InitTable(players ...User)
+	SpinTable()
+	ShuffleTable()
+	NumPlayers() int
+	GetSeating() []User
+	SetSeating([]User)
+	GetCurrentTurn() int
+	SetCurrentTurn(int)
+	GetSeed() int64
+	SetSeed(int64)
+}
+
 var (
-	DefaultSeating     []i.User = []i.User{}
-	DefaultCurrentTurn int      = 0
+	DefaultSeating     []User = []User{}
+	DefaultCurrentTurn int    = 0
 )
 
 type GameTable struct {
-	Seating     []i.User `json:"turns"`
-	CurrentTurn int      `json:"current_turn"`
+	Seating     []User `json:"turns"`
+	CurrentTurn int    `json:"current_turn"`
 	Seed        int64
 }
 
@@ -23,7 +35,7 @@ var DefaultGameTable *GameTable = &GameTable{
 	0,
 }
 
-func (t *GameTable) InitTable(players ...i.User) {
+func (t *GameTable) InitTable(players ...User) {
 	t.SetSeating(players)
 	t.ShuffleTable()
 	t.SetCurrentTurn(0)
@@ -50,11 +62,11 @@ func (t *GameTable) NumPlayers() int {
 	return len(t.GetSeating())
 }
 
-func (t *GameTable) GetSeating() []i.User {
+func (t *GameTable) GetSeating() []User {
 	return t.Seating
 }
 
-func (t *GameTable) SetSeating(players []i.User) {
+func (t *GameTable) SetSeating(players []User) {
 	t.Seating = players
 }
