+++ src/data/revolver.go
@@ -5,6 +5,22 @@ import (
 	"time"
 )
 
+//go:generate mockgen --destination=./../../mocks/gun.go github.com/holy-tech/discord-roulette/src/data Gun
+type Gun interface {
+	ReloadGun(sizeChamber, numBullets int)
+	SpinChamber()
+	ShuffleChamber()
+	Shoot() bool
+	GetNumBulletsLeft() int
+	ChamberSize() int
+	GetSeed() int64
+	SetSeed(seed int64)
+	GetCurrentChamber() int
+	SetCurrentChamber(currChamber int)
+	GetChamber() []bool
+	SetChamber(chamber []bool)
+}
+
 var (
 	DefaultChamber        []bool = []bool{}
 	DefaultNumChamber     int    = 6
